+++ common/common.go
@@ -1,6 +1,6 @@
 package common
 
-//go:generate stringer -type Operation Operation common.go
+//go:generate stringer -type Operation .
 
 /*
 Operation lists the types of operations on each row in a change list.
@@ -48,18 +48,18 @@ type Change struct {
 	// clients should return to the API when looking for a list of changes.
 	Sequence string `json:"sequence,omitempty"`
 	// The LSN when the change was committed.
-	CommitSequence int64 `json:"commitSequence"`
+	CommitSequence uint64 `json:"commitSequence"`
 	// The Postgres LSN when the row changed. This will be interleaved
 	// in the change list because transactions commit at different times.
-	ChangeSequence int64 `json:"changeSequence"`
+	ChangeSequence uint64 `json:"changeSequence"`
 	// The order in which the change happened within the commit. For a transaction
 	// affects multiple rows, changes will be listed in order, and this value
 	// will begin at zero and be incremented by one.
-	CommitIndex int32 `json:"commitIndex"`
+	CommitIndex uint32 `json:"commitIndex"`
 	// The Postgres Transaction ID when this change committed. This may be used
 	// to find a CommitSequence from a particular snapshot. (TXIDs are 32
 	// bits inside Postgres but we may be able to expand this in the future.)
-	TransactionID int64 `json:"txid"`
+	TransactionID uint32 `json:"txid"`
 	// For an insert operation, the columns that are being inserted. For an update,
 	// the new value of the columns
 	NewRow Row `json:"newRow,omitempty"`
